fix crashes on rotation, deletion

+++ app/src/main/java/org/mattvchandler/progressbars/Progress_bars.java
@@ -34,63 +34,66 @@ public class Progress_bars extends AppCompatActivity
     protected void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-
         binding = DataBindingUtil.setContentView(this, R.layout.activity_progress_bars);
         setSupportActionBar(binding.progressBarToolbar);
-
         binding.mainList.addItemDecoration(new DividerItemDecoration(binding.mainList.getContext(), DividerItemDecoration.VERTICAL));
 
-        SQLiteDatabase db = new Progress_bar_DB(this).getWritableDatabase();
+        // only run on 1st creation
+        if(savedInstanceState == null)
+        {
+            SQLiteDatabase db = new Progress_bar_DB(this).getWritableDatabase();
 
-        db.execSQL("DELETE FROM " + Progress_bar_contract.Progress_bar_table.TABLE_NAME);
+            db.execSQL("DELETE FROM " + Progress_bar_contract.Progress_bar_table.TABLE_NAME);
 
-        Progress_bar_data data;
+            Progress_bar_data data;
 
-        data = new Progress_bar_data("ASDF");
-        data.start_time = 1496429958;
-        data.end_time = 1496429958;
-        data.insert(this);
+            data = new Progress_bar_data("ASDF");
+            data.start_time = 1496429958;
+            data.end_time = 1496429958;
+            data.insert(this);
 
-        data = new Progress_bar_data("AC⚡DC");
-        data.start_time = 1451937600;
-        data.end_time = 1483560000;
-        data.insert(this);
+            data = new Progress_bar_data("AC⚡DC");
+            data.start_time = 1451937600;
+            data.end_time = 1483560000;
+            data.insert(this);
 
-        data = new Progress_bar_data("💩");
-        data.start_time = 1388865600;
-        data.end_time = 1388865600;
-        data.insert(this);
+            data = new Progress_bar_data("💩");
+            data.start_time = 1388865600;
+            data.end_time = 1388865600;
+            data.insert(this);
 
-        data = new Progress_bar_data("Queen");
-        data.start_time = 1483261200;
-        data.end_time = 1514797200;
-        data.insert(this);
+            data = new Progress_bar_data("Queen");
+            data.start_time = 1483261200;
+            data.end_time = 1514797200;
+            data.insert(this);
 
-        data = new Progress_bar_data("Led Zepplin");
-        data.start_time = 551476200;
-        data.end_time = 551476200;
-        data.insert(this);
+            data = new Progress_bar_data("Led Zepplin");
+            data.start_time = 551476200;
+            data.end_time = 551476200;
+            data.insert(this);
 
-        data = new Progress_bar_data("Guns N' Roses");
-        data.start_time = 1496429958;
-        data.end_time = 1496429958;
-        data.insert(this);
+            data = new Progress_bar_data("Guns N' Roses");
+            data.start_time = 1496429958;
+            data.end_time = 1496429958;
+            data.insert(this);
 
-        data = new Progress_bar_data("Rolling Stones");
-        data.start_time = 1496429958;
-        data.end_time = 1496429958;
-        data.insert(this);
+            data = new Progress_bar_data("Rolling Stones");
+            data.start_time = 1496429958;
+            data.end_time = 1496429958;
+            data.insert(this);
 
-        data = new Progress_bar_data("Eagles");
-        data.start_time = 1496429958;
-        data.end_time = 1496429958;
-        data.insert(this);
+            data = new Progress_bar_data("Eagles");
+            data.start_time = 1496429958;
+            data.end_time = 1496429958;
+            data.insert(this);
+
+            db.close();
+        }
 
         Cursor cursor = new Progress_bar_DB(this).getReadableDatabase().rawQuery(Progress_bar_contract.Progress_bar_table.SELECT_ALL_ROWS, null);
+        adapter = new Progress_bar_adapter(cursor, this);
 
         binding.mainList.setLayoutManager(new LinearLayoutManager(this));
-        adapter = new Progress_bar_adapter(cursor);
-
         binding.mainList.setAdapter(adapter);
 
         ItemTouchHelper touch_helper = new ItemTouchHelper(new Progress_bar_row_touch_helper_callback(adapter));
